scripts: Nokia 6 desktop setup example
This two lines uses to read the username of the desktop/server which is compiling as well as the PCs name.

This can be changed to manually show up what you want, making it show cm@cm-build-c14 as showed up on Nokia´s kernel version.

Proof for Phandroid to show up on its article.

Change-Id: I576780bae4866242bd67010a982bf2ada852dadd